chore(test): add poll timeouts
