+++ tests/playwright/src/specs/docker-compatibility-smoke.spec.ts
@@ -16,21 +16,16 @@
  * SPDX-License-Identifier: Apache-2.0
  ***********************************************************************/
 
-import type { Page } from '@playwright/test';
-
 import { DockerCompatibilityPage } from '../model/pages/docker-compatibility-page';
 import { PodmanMachineDetails } from '../model/pages/podman-machine-details-page';
-import { PreferencesPage } from '../model/pages/preferences-page';
 import { ResourcesPage } from '../model/pages/resources-page';
 import { SettingsBar } from '../model/pages/settings-bar';
 import { expect as playExpect, test } from '../utility/fixtures';
-import { createPodmanMachineFromCLI } from '../utility/operations';
+import { createPodmanMachineFromCLI, setDockerCompatibilityFeature } from '../utility/operations';
 import { isWindows } from '../utility/platform';
 import { waitForPodmanMachineStartup } from '../utility/wait';
 
 const defaultMachine = 'Podman Machine';
-const DCPreferencesHref = '/preferences/default/preferences.dockerCompatibility';
-const DCSettingsHref = '/preferences/docker-compatibility';
 
 test.beforeAll(async ({ runner, welcomePage, page }) => {
   runner.setVideoAndTraceName('docker-compatibility-e2e');
@@ -41,7 +36,7 @@ test.beforeAll(async ({ runner, welcomePage, page }) => {
 
 test.afterAll(async ({ runner, page }) => {
   if (test.info().status === 'failed') {
-    await setDockerCompatibilityFeature(false, page);
+    await setDockerCompatibilityFeature(page, false);
     await createPodmanMachineFromCLI();
   }
   await runner.close();
@@ -56,11 +51,11 @@ test.describe.serial('Verify docker compatibility feature', { tag: '@smoke' }, (
     await navigationBar.openSettings();
     const settingsBar = new SettingsBar(page);
 
-    const DCLink = settingsBar.getLinkLocatorByHref(DCSettingsHref);
+    const DCLink = settingsBar.getLinkLocatorByHref('/preferences/docker-compatibility');
     await playExpect(DCLink).not.toBeVisible();
 
     //Enable the feature
-    await setDockerCompatibilityFeature(true, page);
+    await setDockerCompatibilityFeature(page, true);
   });
   test('Verify Docker Compatibility page', async ({ page }) => {
     const settingsBar = new SettingsBar(page);
@@ -78,7 +73,9 @@ test.describe.serial('Verify docker compatibility feature', { tag: '@smoke' }, (
     await playExpect(podmanMachineDetails.podmanMachineStatus).toHaveText('OFF', { timeout: 50_000 });
 
     const dockerCompatibilityPage = await settingsBar.openTabPage(DockerCompatibilityPage);
-    await playExpect.poll(async () => await dockerCompatibilityPage.socketIsReachable()).toBeFalsy();
+    await playExpect
+      .poll(async () => await dockerCompatibilityPage.socketIsReachable(), { timeout: 50_000 })
+      .toBeFalsy();
     await playExpect(dockerCompatibilityPage.serverInformationBox).not.toBeVisible();
 
     await settingsBar.openTabPage(ResourcesPage);
@@ -87,48 +84,12 @@ test.describe.serial('Verify docker compatibility feature', { tag: '@smoke' }, (
     await playExpect(podmanMachineDetails.podmanMachineStatus).toHaveText('RUNNING', { timeout: 50_000 });
 
     await settingsBar.openTabPage(DockerCompatibilityPage);
-    await playExpect.poll(async () => await dockerCompatibilityPage.socketIsReachable()).toBeTruthy();
+    await playExpect
+      .poll(async () => await dockerCompatibilityPage.socketIsReachable(), { timeout: 50_000 })
+      .toBeTruthy();
     await playExpect(dockerCompatibilityPage.serverInformationBox).toBeVisible();
   });
   test('Disable docker compatibility', async ({ page }) => {
-    await setDockerCompatibilityFeature(false, page);
+    await setDockerCompatibilityFeature(page, false);
   });
 });
-
-async function setDockerCompatibilityFeature(enable: boolean, page: Page): Promise<void> {
-  //Open the preferences bar and verify DC preferences page
-  const settingsBar = new SettingsBar(page);
-  await settingsBar.expandPreferencesTab();
-
-  const DCPreferencesLink = settingsBar.getLinkLocatorByHref(DCPreferencesHref);
-  await playExpect(DCPreferencesLink).toBeVisible();
-  await DCPreferencesLink.click();
-  const DCPreferencesPage = new PreferencesPage(page);
-
-  await playExpect(DCPreferencesPage.heading).toBeVisible();
-  const experimentalTitle = DCPreferencesPage.content.getByText('Docker Compatibility', { exact: true });
-  await playExpect(experimentalTitle).toBeVisible();
-
-  //Set the feature
-  const dockerCompatibilityCheckbox = DCPreferencesPage.content.getByRole('checkbox', {
-    name: 'Enable the section for Docker compatibility.',
-  });
-  await playExpect(dockerCompatibilityCheckbox).toBeVisible();
-  const isEnabled = await dockerCompatibilityCheckbox.isChecked();
-  if (isEnabled !== enable) {
-    await dockerCompatibilityCheckbox.locator('..').setChecked(enable);
-    const isEnabled = await dockerCompatibilityCheckbox.isChecked();
-    playExpect(isEnabled).toEqual(enable);
-  }
-
-  //Verify the main docker compatibility page (dis)appeared
-  const DCSettingsLink = settingsBar.getLinkLocatorByHref(DCSettingsHref);
-  if (enable) {
-    await playExpect(DCSettingsLink).toBeVisible();
-  } else {
-    await playExpect(DCSettingsLink).not.toBeVisible();
-  }
-
-  //Close the preferences bar
-  await settingsBar.expandPreferencesTab();
-}
